backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.228
opencv-python==4.8.1.78
pillow>=10.0.0
numpy>=1.24.0
pdf2image==1.16.3
